Move preview components inside package
