+++ android/sample/src/main/kotlin/com/gu/source/MainActivity.kt
@@ -21,6 +21,7 @@ import com.gu.source.presets.palette.Neutral10
 import com.gu.source.presets.palette.Neutral100
 import com.gu.source.presets.palette.Neutral97
 import com.gu.source.presets.typography.HeadlineMedium20
+import com.gu.source.previews.*
 import com.gu.source.utils.PreviewPhoneBothMode
 import com.gu.source.utils.plus
 import kotlinx.coroutines.launch
